move to directory with index page
