+++ src/pages/reviewed-preprints/[...doi]/index.module.scss
@@ -1,5 +1,5 @@
-@use "../../sass/settings";
-@use "../../sass/layout";
+@use "../../../sass/settings";
+@use "../../../sass/layout";
 
 .article-page {
   @include settings.padding();
